feat: Separado porcamadas e criado metodo para criação do user no banco de dados
