+++ API/ECOMMERCE.API/ECOMMERCE.API.sln
@@ -5,6 +5,10 @@ VisualStudioVersion = 17.14.36401.2 d17.14
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ECOMMERCE.API", "ECOMMERCE.API\ECOMMERCE.API.csproj", "{3013DC6E-B4F5-4DA0-885D-2732B2652423}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ECOMMERCE.CORE", "ECOMMERCE.CORE\ECOMMERCE.CORE.csproj", "{B7858626-158F-4B2D-BF08-5225BAD652FF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ECOMMERCE.DATA", "ECOMMERCE.DATA\ECOMMERCE.DATA.csproj", "{884865AE-D21D-4367-BF79-45CF265059F5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +19,14 @@ Global
 		{3013DC6E-B4F5-4DA0-885D-2732B2652423}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3013DC6E-B4F5-4DA0-885D-2732B2652423}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3013DC6E-B4F5-4DA0-885D-2732B2652423}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B7858626-158F-4B2D-BF08-5225BAD652FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B7858626-158F-4B2D-BF08-5225BAD652FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B7858626-158F-4B2D-BF08-5225BAD652FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B7858626-158F-4B2D-BF08-5225BAD652FF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{884865AE-D21D-4367-BF79-45CF265059F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{884865AE-D21D-4367-BF79-45CF265059F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{884865AE-D21D-4367-BF79-45CF265059F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{884865AE-D21D-4367-BF79-45CF265059F5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
