+++ API/ECOMMERCE.API/ECOMMERCE.API/Controllers/UserController.cs
@@ -1,0 +1,32 @@
+using System.Security.Claims;
+using ECOMMERCE.CORE.DTO.User;
+using ECOMMERCE.CORE.Interfaces;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+
+namespace ECOMMERCE.API.Controllers;
+
+[ApiController]
+[Route("api/[controller]")]
+public class UserController : Controller
+{
+    private readonly IUserService _userService;
+
+    public UserController(IUserService userService)
+    {
+        _userService = userService;
+    }
+
+    // [Authorize]
+    [HttpGet]
+    public async Task<IActionResult> VerifyUser()
+    {
+        string keycloakId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? User.FindFirst("sub")?.Value;
+        
+        GetUserDTO userLogin = _userService.VerifyUser(keycloakId);
+
+        return Ok(userLogin);
+    }
+    
+    
+}
