--- API/ECOMMERCE.API/ECOMMERCE.API/DTO/Authorization/PostTokenDTO.cs
@@ -1,7 +1,0 @@
-namespace ECOMMERCE.API.DTO.Authorization;
-
-public class PostTokenDTO
-{
-    public string Usuario { get; set; }
-    public string Senha { get; set; }
-}
