--- API/ECOMMERCE.API/ECOMMERCE.API/DTO/Product/CreateProductDTO.cs
@@ -1,11 +1,0 @@
-namespace ECOMMERCE.API.DTO.Product;
-
-public class CreateProductDTO
-{
-    public string Name { get; set; }
-    public string Description { get; set; }
-    public Decimal Price { get; set; }
-    public string ImageUrl { get; set; }
-    public int Stock { get; set; }
-    public string? IdPai { get; set; }
-}
