--- API/ECOMMERCE.API/ECOMMERCE.API/DTO/User/CreateAddressDTO.cs
@@ -1,9 +1,0 @@
-﻿namespace ECOMMERCE.API.DTO.User
-{
-    public class CreateAddressDTO
-    {
-        public string Street { get; set; }
-        public string Number { get; set; }
-        public string CEP { get; set; }
-    }
-}
