--- API/ECOMMERCE.API/ECOMMERCE.API/DTO/User/GetAddressDTO.cs
@@ -1,9 +1,0 @@
-﻿namespace ECOMMERCE.API.DTO.User
-{
-    public class GetAddressDTO
-    {
-        public string Street { get; set; }
-        public string Number { get; set; }
-        public string CEP { get; set; }
-    }
-}
