+++ API/ECOMMERCE.API/ECOMMERCE.API/ECOMMERCE.API.csproj
@@ -8,20 +8,22 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="8.0.19" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.8" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.8">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
 
       <PackageReference Include="IdentityModel.AspNetCore" Version="4.0.0" />
       <PackageReference Include="Keycloak.AuthServices.Authentication" Version="1.2.1" />
       <PackageReference Include="Keycloak.AuthServices.Authorization" Version="1.2.1" />
       <PackageReference Include="Keycloak.Net" Version="1.0.18" />
+      <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.8">
+        <PrivateAssets>all</PrivateAssets>
+        <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+      </PackageReference>
       <PackageReference Include="Refit.HttpClientFactory" Version="8.0.0" />
-    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="9.0.4" />
-    <PackageReference Include="Pomelo.EntityFrameworkCore.MySql" Version="9.0.0" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.6.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\ECOMMERCE.CORE\ECOMMERCE.CORE.csproj" />
+    <ProjectReference Include="..\ECOMMERCE.DATA\ECOMMERCE.DATA.csproj" />
   </ItemGroup>
 
 </Project>
