--- API/ECOMMERCE.API/ECOMMERCE.API/Entity/Address.cs
@@ -1,14 +1,0 @@
-﻿namespace ECOMMERCE.API.Entity
-{
-    public class Address
-    {
-        public Guid Id { get; set; }
-        public string Street { get; set; }
-        public string number { get; set; }
-        public string CEP { get; set; }
-        public Address()
-        {
-            
-        }
-    }
-}
