--- API/ECOMMERCE.API/ECOMMERCE.API/Entity/Sale.cs
@@ -1,8 +1,0 @@
-namespace ECOMMERCE.API.Entity;
-
-public class Sale
-{
-    public Guid Id { get; set; }
-    public Guid UserId { get; set; }
-    public User User { get; set; }
-}
