--- API/ECOMMERCE.API/ECOMMERCE.API/Entity/User.cs
@@ -1,13 +1,0 @@
-﻿namespace ECOMMERCE.API.Entity
-{
-    public class User
-    {
-        public Guid Id { get; set; }
-        public string KeycloakId { get; set; }
-        public Address? Address { get; set; }
-        public User()
-        {
-    
-        }
-    }
-}
