--- API/ECOMMERCE.API/ECOMMERCE.API/Interfaces/IAuthorizationService.cs
@@ -1,8 +1,0 @@
-using ECOMMERCE.API.DTO.Authorization;
-
-namespace ECOMMERCE.API.Interfaces;
-
-public interface IAuthorizationService
-{
-    public Task<string> PostToken(PostTokenDTO dto);
-}
