--- API/ECOMMERCE.API/ECOMMERCE.API/Interfaces/IUserService.cs
@@ -1,8 +1,0 @@
-using ECOMMERCE.API.Entity;
-
-namespace ECOMMERCE.API.Interfaces;
-
-public interface IUserService
-{
-    public User VerifyUser(string keycloakId);
-}
