+++ API/ECOMMERCE.API/ECOMMERCE.CORE/DTO/Authorization/PostTokenDTO.cs
@@ -1,0 +1,7 @@
+namespace ECOMMERCE.CORE.DTO.Authorization;
+
+public class PostTokenDTO
+{
+    public string Usuario { get; set; }
+    public string Senha { get; set; }
+}
