+++ API/ECOMMERCE.API/ECOMMERCE.CORE/DTO/Product/CreateProductDTO.cs
@@ -1,0 +1,11 @@
+namespace ECOMMERCE.CORE.DTO.Product;
+
+public class CreateProductDTO
+{
+    public string Name { get; set; }
+    public string Description { get; set; }
+    public Decimal Price { get; set; }
+    public string ImageUrl { get; set; }
+    public int Stock { get; set; }
+    public string? IdPai { get; set; }
+}
