+++ API/ECOMMERCE.API/ECOMMERCE.CORE/DTO/User/CreateAddressDTO.cs
@@ -1,0 +1,9 @@
+﻿namespace ECOMMERCE.CORE.DTO.User
+{
+    public class CreateAddressDTO
+    {
+        public string Street { get; set; }
+        public string Number { get; set; }
+        public string CEP { get; set; }
+    }
+}
