+++ API/ECOMMERCE.API/ECOMMERCE.CORE/Entity/Address.cs
@@ -1,0 +1,14 @@
+﻿namespace ECOMMERCE.CORE.Entity
+{
+    public class Address
+    {
+        public Guid Id { get; set; }
+        public string Street { get; set; }
+        public string number { get; set; }
+        public string CEP { get; set; }
+        public Address()
+        {
+            
+        }
+    }
+}
