+++ API/ECOMMERCE.API/ECOMMERCE.CORE/Entity/Product.cs
@@ -1,0 +1,14 @@
+namespace ECOMMERCE.CORE.Entity;
+
+public class Product
+{
+    public Guid Id { get; set; }
+    public string Name { get; set; } 
+    public string Description { get; set; } 
+    public Decimal Price { get; set; }
+    public string ImageUrl { get; set; }
+    public int Stock { get; set; }
+    public Guid? IdPai { get; set; }
+    
+    
+}
