+++ API/ECOMMERCE.API/ECOMMERCE.CORE/Entity/Sale.cs
@@ -1,0 +1,8 @@
+namespace ECOMMERCE.CORE.Entity;
+
+public class Sale
+{
+    public Guid Id { get; set; }
+    public Guid UserId { get; set; }
+    public User User { get; set; }
+}
