+++ API/ECOMMERCE.API/ECOMMERCE.CORE/Entity/User.cs
@@ -1,0 +1,14 @@
+﻿namespace ECOMMERCE.CORE.Entity
+{
+    public class User
+    {
+        public Guid Id { get; set; }
+        public string KeycloakId { get; set; }
+        public Guid? AddressId { get; set; }
+        public Address? Address { get; set; }
+        public User()
+        {
+    
+        }
+    }
+}
