+++ API/ECOMMERCE.API/ECOMMERCE.CORE/Interfaces/IAuthorizationService.cs
@@ -1,0 +1,8 @@
+using ECOMMERCE.CORE.DTO.Authorization;
+
+namespace ECOMMERCE.CORE.Interfaces;
+
+public interface IAuthorizationService
+{
+    public Task<string> PostToken(PostTokenDTO dto);
+}
