+++ API/ECOMMERCE.API/ECOMMERCE.CORE/Interfaces/IUserRepository.cs
@@ -1,0 +1,10 @@
+using ECOMMERCE.CORE.Entity;
+
+namespace ECOMMERCE.CORE.Interfaces;
+
+public interface IUserRepository
+{
+    public User GetUserByKeycloakId(string keycloakId);
+    
+    public User CreateUser(string keycloakId);
+}
