+++ API/ECOMMERCE.API/ECOMMERCE.CORE/Interfaces/IUserService.cs
@@ -1,0 +1,9 @@
+using ECOMMERCE.CORE.DTO.User;
+using ECOMMERCE.CORE.Entity;
+
+namespace ECOMMERCE.CORE.Interfaces;
+
+public interface IUserService
+{
+    public GetUserDTO VerifyUser(string keycloakId);
+}
