+++ API/ECOMMERCE.API/ECOMMERCE.CORE/Services/UserService.cs
@@ -1,0 +1,31 @@
+using ECOMMERCE.CORE.DTO.User;
+using ECOMMERCE.CORE.Entity;
+using ECOMMERCE.CORE.Interfaces;
+
+namespace ECOMMERCE.CORE.Services;
+
+public class UserService : IUserService
+{
+    private readonly IUserRepository _userRepository;
+
+    public UserService(IUserRepository userRepository)
+    {
+        _userRepository =  userRepository;
+    }
+    
+    public GetUserDTO VerifyUser(string keycloakId)
+    {
+        User user = _userRepository.GetUserByKeycloakId(keycloakId);
+
+        if (user == null)
+            user = _userRepository.CreateUser(keycloakId);
+
+        GetUserDTO userDTO = new GetUserDTO()
+        {
+            Id = user.Id,
+            keycloakId = user.KeycloakId
+        };
+        
+        return userDTO;
+    }
+}
