+++ API/ECOMMERCE.API/ECOMMERCE.DATA/Repositories/UserRepository.cs
@@ -1,0 +1,34 @@
+using System.Security.Cryptography.X509Certificates;
+using ECOMMERCE.CORE.Entity;
+using ECOMMERCE.CORE.Interfaces;
+using ECOMMERCE.DATA.Data;
+
+namespace ECOMMERCE.DATA.Repositories;
+
+public class UserRepository : IUserRepository
+{
+    private readonly EcommerceDbContext _ecommerceDbContext;
+
+    public UserRepository(EcommerceDbContext ecommerceDbContext)
+    {
+        _ecommerceDbContext = ecommerceDbContext;
+    }
+    public User GetUserByKeycloakId(string keycloakId)
+    {
+        User user = _ecommerceDbContext.Users.FirstOrDefault(user => user.KeycloakId == keycloakId);
+        
+        return user;
+    }
+
+    public User CreateUser(string keycloakId)
+    {
+        User user = new User();
+        user.KeycloakId = keycloakId;
+
+        _ecommerceDbContext.Users.Add(user);
+        
+        _ecommerceDbContext.SaveChanges();
+        
+        return user;
+    }
+}
